apps/backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
sqlmodel
asyncpg
psycopg2-binary
alembic
pydantic-settings
python-multipart
httpx
firecrawl-py
playwright
boto3
openai
python-dotenv
psycopg2-binary
trafilatura==1.6.0
pymupdf==1.23.8
beautifulsoup4
bleach
